pyproject.toml
[project]
name = "whatsmcp"
version = "1.0.0"
description = "WhatsApp bot powered by Claude Code SDK via WhatsGateway"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "claude-agent-sdk>=0.1.0",
    "structlog>=24.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
]

[project.scripts]
whatsmcp = "whatsmcp.main:run"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["whatsmcp*"]
